added configuration parameter to disable tracking of statistics not required for block validation

+++ libraries/blockchain/chain_database.cpp
@@ -868,8 +868,11 @@ namespace bts { namespace blockchain {
           delegate_info.last_block_num_produced = produced_block.block_num;
           pending_state->store_account_record( *delegate_record );
 
-          const slot_record slot( produced_block.timestamp, delegate_id, produced_block.id() );
-          pending_state->store_slot_record( slot );
+          if( _track_stats )
+          {
+             const slot_record slot( produced_block.timestamp, delegate_id, produced_block.id() );
+             pending_state->store_slot_record( slot );
+          }
 
           /* Update production info for missing delegates */
 
@@ -893,7 +896,8 @@ namespace bts { namespace blockchain {
               delegate_record->delegate_info->blocks_missed += 1;
               pending_state->store_account_record( *delegate_record );
 
-              pending_state->store_slot_record( slot_record( block_timestamp, delegate_id )  );
+             if( _track_stats )
+                 pending_state->store_slot_record( slot_record( block_timestamp, delegate_id )  );
           }
 
           /* Update required confirmation count */
@@ -949,11 +953,12 @@ namespace bts { namespace blockchain {
            market_engine engine( pending_state, *this );
            if( engine.execute( market_pair.first, market_pair.second, timestamp ) )
            {
-              market_transactions.insert( market_transactions.end(), engine._market_transactions.begin(), engine._market_transactions.end() );
+              if( _track_stats )
+                 market_transactions.insert( market_transactions.end(), engine._market_transactions.begin(), engine._market_transactions.end() );
            }
         }
-
-        pending_state->set_market_transactions( std::move( market_transactions ) );
+        if( _track_stats )
+           pending_state->set_market_transactions( std::move( market_transactions ) );
       } FC_CAPTURE_AND_RETHROW() }
 
       /**
@@ -1975,7 +1980,8 @@ namespace bts { namespace blockchain {
     { try {
         ilog("@n storing edge in chain DB: ${e}", ("e", edge));
         my->_edge_index.store( edge.index_key(), edge._id );
-        my->_reverse_edge_index.store( edge.reverse_index_key(), edge._id );
+        if( my->_track_stats )
+           my->_reverse_edge_index.store( edge.reverse_index_key(), edge._id );
         my->_object_db.store( edge._id, edge );
     } FC_CAPTURE_AND_RETHROW( (edge) ) }
 
@@ -2041,6 +2047,7 @@ namespace bts { namespace blockchain {
 
    otransaction_record chain_database::get_transaction( const transaction_id_type& trx_id, bool exact )const
    { try {
+      FC_ASSERT( my->_track_stats );
       auto trx_rec = my->_id_to_transaction_record_db.fetch_optional( trx_id );
       if( trx_rec || exact )
       {
@@ -2069,13 +2076,15 @@ namespace bts { namespace blockchain {
    { try {
       if( record_to_store.trx.operations.size() == 0 )
       {
-        my->_id_to_transaction_record_db.remove( record_id );
+        if( my->_track_stats )
+           my->_id_to_transaction_record_db.remove( record_id );
         my->_unique_transactions[record_to_store.trx.expiration].erase( record_to_store.trx.digest(my->_chain_id) );
       }
       else
       {
         FC_ASSERT( record_id == record_to_store.trx.id() );
-        my->_id_to_transaction_record_db.store( record_id, record_to_store );
+        if( my->_track_stats )
+           my->_id_to_transaction_record_db.store( record_id, record_to_store );
         if( record_to_store.trx.expiration > this->now() )
            FC_ASSERT( my->_unique_transactions[record_to_store.trx.expiration].insert( record_to_store.trx.digest(my->_chain_id) ).second );
       }
@@ -2470,6 +2479,7 @@ namespace bts { namespace blockchain {
     vector<slot_record> chain_database::get_delegate_slot_records( const account_id_type& delegate_id,
                                                                    int64_t start_block_num, uint32_t count )const
     {
+        FC_ASSERT( my->_track_stats, "index of slot records is disabled" );
         FC_ASSERT( count > 0 );
         if( start_block_num < 0 )
             start_block_num = int64_t( get_head_block_num() ) + start_block_num;
@@ -3143,19 +3153,22 @@ namespace bts { namespace blockchain {
 
    void chain_database::store_slot_record( const slot_record& r )
    {
-       if( r.is_null() )
-           my->_slot_record_db.remove( r.start_time );
-       else
-           my->_slot_record_db.store( r.start_time, r );
+     if( !my->_track_stats ) return;
+     if( r.is_null() )
+         my->_slot_record_db.remove( r.start_time );
+     else
+         my->_slot_record_db.store( r.start_time, r );
    }
 
    oslot_record chain_database::get_slot_record( const time_point_sec& start_time )const
    {
+     FC_ASSERT( my->_track_stats );
      return my->_slot_record_db.fetch_optional( start_time );
    }
 
    void chain_database::store_market_history_record(const market_history_key& key, const market_history_record& record)
    {
+     if( !my->_track_stats ) return;
      if( record.volume == 0 )
        my->_market_history_db.remove( key );
      else
@@ -3164,6 +3177,7 @@ namespace bts { namespace blockchain {
 
    omarket_history_record chain_database::get_market_history_record(const market_history_key& key) const
    {
+     FC_ASSERT( my->_track_stats );
      return my->_market_history_db.fetch_optional( key );
    }
 
@@ -3250,6 +3264,7 @@ namespace bts { namespace blockchain {
 
    void chain_database::set_market_transactions( vector<market_transaction> trxs )
    {
+      FC_ASSERT( my->_track_stats );
       if( trxs.size() == 0 )
       {
          my->_market_transactions_db.remove( get_head_block_num()+1 );
@@ -3262,6 +3277,7 @@ namespace bts { namespace blockchain {
 
    vector<market_transaction> chain_database::get_market_transactions( uint32_t block_num  )const
    {
+      FC_ASSERT( my->_track_stats );
       auto tmp = my->_market_transactions_db.fetch_optional(block_num);
       if( tmp ) return *tmp;
       return vector<market_transaction>();
@@ -3563,6 +3579,9 @@ namespace bts { namespace blockchain {
 
    void chain_database::store_burn_record( const burn_record& br )
    {
+      if( !my->_track_stats ) 
+         return;
+
       if( br.is_null() )
       {
          my->_burn_db.remove( br );
@@ -3573,6 +3592,7 @@ namespace bts { namespace blockchain {
 
    oburn_record chain_database::fetch_burn_record( const burn_record_key& key )const
    {
+      FC_ASSERT( my->_track_stats );
       auto oval = my->_burn_db.fetch_optional( key );
       if( oval )
          return burn_record( key, *oval );
@@ -3580,6 +3600,7 @@ namespace bts { namespace blockchain {
    }
    vector<burn_record> chain_database::fetch_burn_records( const string& account_name )const
    { try {
+      FC_ASSERT( my->_track_stats );
       vector<burn_record> results;
       auto opt_account_record = get_account_record( account_name );
       FC_ASSERT( opt_account_record.valid() );
@@ -3766,10 +3787,12 @@ namespace bts { namespace blockchain {
    }
    void     chain_database::index_transaction( const address& addr, const transaction_id_type& trx_id )
    {
-      my->_address_to_trx_index.store( std::make_pair(addr,trx_id), char(0) );
+      if( my->_track_stats )
+         my->_address_to_trx_index.store( std::make_pair(addr,trx_id), char(0) );
    }
    vector<transaction_record>  chain_database::fetch_address_transactions( const address& addr )
    {
+      FC_ASSERT( my->_track_stats );
       vector<transaction_record> results;
       auto itr = my->_address_to_trx_index.lower_bound( std::make_pair(addr, transaction_id_type()) );
       while( itr.valid() )
@@ -3784,6 +3807,10 @@ namespace bts { namespace blockchain {
          ++itr;
       }
       return results;
+   }
+   void chain_database::track_chain_statistics( bool status )
+   {
+      my->_track_stats = status;
    }
 
 } } // bts::blockchain
